+++ tvtarian/crawlers/shared/const.py
@@ -1,0 +1,5 @@
+ANOTHER_SKY_ID = 1
+JOHNETSU_ID = 2
+PROFESSIONAL_ID = 3
+SEVEN_RULE_ID = 4
+MUSIC_STATION_ID = 5
